unify numerical type to f32

+++ xprite/src/algorithms/line.rs
@@ -104,22 +104,22 @@ pub fn snapped_line(is_45: bool, start: &Pixel, stop: &Pixel) -> Vec<Pixel> {
     ret
 }
 
-pub fn bresenham(start: &Point2D<u32>, stop: &Point2D<u32>) -> Vec<Pixel> {
+pub fn bresenham(start: &Point2D<f32>, stop: &Point2D<f32>) -> Vec<Pixel> {
     let mut ret = Vec::new();
-    let mut x0 = start.x as i32;
-    let mut y0 = start.y as i32;
-    let x1 = stop.x as i32;
-    let y1 = stop.y as i32;
+    let mut x0 = start.x;
+    let mut y0 = start.y;
+    let x1 = stop.x;
+    let y1 = stop.y;
 
     let dx = (x1-x0).abs();
-    let sx = if x0<x1 {1} else {-1};
+    let sx = if x0<x1 {1.} else {-1.};
     let dy = -(y1-y0).abs();
-    let sy = if y0<y1 {1} else {-1};
+    let sy = if y0<y1 {1.} else {-1.};
     let mut err = dx+dy; /* error value e_xy */
     loop {
         ret.push(pixel!(x0, y0));
         if x0==x1 && y0==y1 { break; }
-        let e2 = 2 * err;
+        let e2 = 2. * err;
         if e2 >= dy { err += dy; x0 += sx; } /* e_xy+e_x > 0 */
         if e2 <= dx { err += dx; y0 += sy; } /* e_xy+e_y < 0 */
     }
@@ -132,7 +132,7 @@ mod tests {
     use super::*;
     #[test]
     fn test_adjust() {
-        let adjusted_end = snapped_line(true, &pixel!(0, 0), &pixel!(10, 9));
+        let adjusted_end = snapped_line(true, &pixel!(0., 0.), &pixel!(10., 9.));
         println!("{:?}", adjusted_end);
     }
 }
